# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Dollar CARE'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/v1/appointment-store:
    post:
      summary: 'Store Appointment'
      operationId: storeAppointment
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Appointment booked successfully!'
                  status: true
                  data:
                    user_id: 1
                    doctor_id: 4
                    clinic_id: 7
                    appointment_id: '000000'
                    appointment_date: '2021-08-02'
                    appointment_time: '11:00 AM'
                    appointment_status: Done
                    booking_time: '2021-08-02 12:00:00'
                    clinic_name: 'Life Medical'
                    clinic_address: 'Park Street, Mullick Bazar, Beniapukur, Kolkata, West Bengal, India'
                    clinic_phone: '7894561230'
                    updated_at: '2021-08-02T12:00:00.000000Z'
                    created_at: '2021-08-02T12:00:00.000000Z'
                    id: 1
                    user:
                      id: 1
                      name: 'Shilpi Chaki'
                      email: shilpi@mailinator.com
                      email_verified_at: null
                      profile_picture: user/5HiPk9oN9cQCNNKdzmBgvNLHSL8u7bbHncdrPE91.png
                      phone: '7894561230'
                      year_of_experience: '4'
                      license_number: UPS74856963
                      location: Kolkata
                      gender: female
                      age: '2001-01-30'
                      status: 1
                      fcm_token: 'fTcKacaIV3mJaU4VbJ_4Ib:APA91bHgCcot6OOpAlysNgDKq54K9PnU6PjDM3tHdiWFm5KpPXDnVVfcfTJV2C4Q6wKk056fIp6zgOLbEU1DhAvN5SQqZGg0ew7qjKN_FAXyd9ORER4vMhjf1qLc9pY32gI7ZGIJLAaI'
                      created_at: '2021-07-25T08:53:27.000000Z'
                      updated_at: '2021-07-25T08:53:27.000000Z'
                      deleted_at: null
                properties:
                  message:
                    type: string
                    example: 'Appointment booked successfully!'
                  status:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user_id:
                        type: integer
                        example: 1
                      doctor_id:
                        type: integer
                        example: 4
                      clinic_id:
                        type: integer
                        example: 7
                      appointment_id:
                        type: string
                        example: '000000'
                      appointment_date:
                        type: string
                        example: '2021-08-02'
                      appointment_time:
                        type: string
                        example: '11:00 AM'
                      appointment_status:
                        type: string
                        example: Done
                      booking_time:
                        type: string
                        example: '2021-08-02 12:00:00'
                      clinic_name:
                        type: string
                        example: 'Life Medical'
                      clinic_address:
                        type: string
                        example: 'Park Street, Mullick Bazar, Beniapukur, Kolkata, West Bengal, India'
                      clinic_phone:
                        type: string
                        example: '7894561230'
                      updated_at:
                        type: string
                        example: '2021-08-02T12:00:00.000000Z'
                      created_at:
                        type: string
                        example: '2021-08-02T12:00:00.000000Z'
                      id:
                        type: integer
                        example: 1
                      user:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          name:
                            type: string
                            example: 'Shilpi Chaki'
                          email:
                            type: string
                            example: shilpi@mailinator.com
                          email_verified_at:
                            type: string
                            example: null
                          profile_picture:
                            type: string
                            example: user/5HiPk9oN9cQCNNKdzmBgvNLHSL8u7bbHncdrPE91.png
                          phone:
                            type: string
                            example: '7894561230'
                          year_of_experience:
                            type: string
                            example: '4'
                          license_number:
                            type: string
                            example: UPS74856963
                          location:
                            type: string
                            example: Kolkata
                          gender:
                            type: string
                            example: female
                          age:
                            type: string
                            example: '2001-01-30'
                          status:
                            type: integer
                            example: 1
                          fcm_token:
                            type: string
                            example: 'fTcKacaIV3mJaU4VbJ_4Ib:APA91bHgCcot6OOpAlysNgDKq54K9PnU6PjDM3tHdiWFm5KpPXDnVVfcfTJV2C4Q6wKk056fIp6zgOLbEU1DhAvN5SQqZGg0ew7qjKN_FAXyd9ORER4vMhjf1qLc9pY32gI7ZGIJLAaI'
                          created_at:
                            type: string
                            example: '2021-07-25T08:53:27.000000Z'
                          updated_at:
                            type: string
                            example: '2021-07-25T08:53:27.000000Z'
                          deleted_at:
                            type: string
                            example: null
      tags:
        - 'Patient Booking'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clinic_id:
                  type: string
                  description: ''
                  example: quia
                appointment_date:
                  type: string
                  description: ''
                  example: unde
                appointment_time:
                  type: string
                  description: ''
                  example: et
              required:
                - clinic_id
                - appointment_date
                - appointment_time
      security: []
...